feat: implement routing for health check and subscriptions, add database migration script
