+++ src/lib.rs
@@ -1,31 +1,3 @@
-use std::net::TcpListener;
-
-use actix_web::{dev::Server, web, App, HttpRequest, HttpResponse, HttpServer};
-use serde::Deserialize;
-
-async fn health_check(_req: HttpRequest) -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-#[derive(Deserialize)]
-struct FormData {
-    email: String,
-    name: String,
-}
-
-// Form already verify if data from URL will deserialze correctly on FormData struct
-async fn subscribe(form: web::Form<FormData>) -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(move || {
-        App::new()
-            .route("/health_check", web::get().to(health_check))
-            .route("/subscriptions", web::post().to(subscribe))
-    })
-    .listen(listener)?
-    .run();
-
-    Ok(server)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
