+++ src/routes/health_check.rs
@@ -1,0 +1,5 @@
+use actix_web::{HttpRequest, HttpResponse};
+
+pub async fn health_check(_req: HttpRequest) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
