+++ src/routes/subscriptions.rs
@@ -1,0 +1,13 @@
+use actix_web::{web, HttpResponse};
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct FormData {
+    email: String,
+    name: String,
+}
+
+// Form already verify if data from URL will deserialze correctly on FormData struct
+pub async fn subscribe(form: web::Form<FormData>) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
